requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.1
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
Django==5.1.5
executing==2.2.0
ffmpeg-python==0.2.0
filelock==3.17.0
ftfy==6.3.1
future==1.0.0
gallery_dl==1.28.5
humanize==4.11.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
library==3.0.49
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mutagen==1.47.0
natsort==8.4.0
npm==0.1.1
optional-django==0.1.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg2==2.9.10
pure_eval==0.2.3
puremagic==1.28
pycryptodomex==3.21.0
Pygments==2.19.1
pyreadline3==3.5.4
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-mpv-jsonipc==1.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
screeninfo==0.8.1
six==1.17.0
soupsieve==2.6
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tinytag==2.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
virtualenv==20.29.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
yt-dlp==2025.1.26

djangorestframework~=3.15.2
graphene-django~=3.2.2
pandas~=2.2.3
graphene~=3.4.3